fix: resolve Int to Long type mismatches in UI and DB mappers for amounts/balances

+++ core/domain/src/commonMain/kotlin/com/kazemieh/domain/repository/TransactionRepository.kt
@@ -88,10 +88,10 @@ interface TransactionRepository {
     suspend fun saveRecentSearch(query: String)
     suspend fun deleteRecentSearch(query: String)
 
-    suspend fun updateCategoryPositions(positions: Map<Long, Long>)
-    suspend fun updateSourcePositions(positions: Map<Long, Long>)
-    suspend fun updateTagPositions(positions: Map<Long, Long>)
-    suspend fun updatePersonPositions(positions: Map<Long, Long>)
+    suspend fun updateCategoryPositions(positions: Map<Long, Int>)
+    suspend fun updateSourcePositions(positions: Map<Long, Int>)
+    suspend fun updateTagPositions(positions: Map<Long, Int>)
+    suspend fun updatePersonPositions(positions: Map<Long, Int>)
 
     suspend fun getTransactionAmountRange(): Pair<Long, Long>
